Split GCS storage backend modules
